% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexkart.R
\docType{package}
\name{hexkart-package}
\alias{hexkart}
\alias{hexkart-package}
\title{Kart Simulation and Capturing Tools}
\description{
This package provides tools to render and capture kart training data using
'HTML'  widgets  and 'Shiny' applications.
}
\details{
This package can be used to: Collect kart playing styles, create test cases
for image classification and render trained models over an interactive
kart simulation.
}
\author{
\strong{Maintainer}: Javier Luraschi \email{javier@rstudio.com}

Authors:
\itemize{
  \item  three.js authors
  \item Rick Bruns \email{rick@nintendomaps.com} (Circuit Texture)
  \item  khinsider.com (Circuit Audio)
}

}
