% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{hexkart_capture}
\alias{hexkart_capture}
\title{Captures Training}
\usage{
hexkart_capture(target_path = "capture", width = 80, height = 40)
}
\arguments{
\item{target_path}{Destination path for captured images.}

\item{width}{Width of captured image.}

\item{height}{Height of captured image.

#' @examples 

library(hexkart)
if (interactive()) {
  hexkart_capture()
}}
}
\description{
Creates an ShinyGadget to capture training
}
