% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{hexkart_control}
\alias{hexkart_control}
\title{Controls Simulation}
\usage{
hexkart_control(direction, width = 80, height = 40)
}
\arguments{
\item{direction}{A function \code{function(image, direction) {}} that provides
control for this kart as a string with valid values: \code{"left"}, 
\code{"forward"} or \code{"right"}.}

\item{width}{Width of captured image.}
}
\description{
Creates an ShinyGadget to control kart in simulation.
}
