% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{hexkart_play}
\alias{hexkart_play}
\title{Play Simulation}
\usage{
hexkart_play(width = NULL, height = NULL, circuit = 1)
}
\arguments{
\item{width}{While capturing, width of capture image.}

\item{height}{While capturing, height of capture image.}

\item{circuit}{The circuit index, valid values: 1 or 2.}
}
\description{
Creates an HTMLWidget to play the simulation.
}
\examples{

library(hexkart)
show_pixels(
  round(runif(400, 0, 1)),
  grid = c(40, 10),
  size = c(800, 200),
  params = list(fill = list(color = "#FF3388"))
)

}
