% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{kartsim_capture}
\alias{kartsim_capture}
\title{Captures Training}
\usage{
kartsim_capture(target_path = "capture", width = 32, height = 32,
  circuit = 1)
}
\arguments{
\item{target_path}{Destination path for captured images.}

\item{width}{Width of captured image.}

\item{height}{Height of captured image.}

\item{circuit}{The circuit index, valid values: 1 or 2.

#' @examples 

library(kartsim)
if (interactive()) {
  kartsim_capture()
}}
}
\description{
Creates an ShinyGadget to capture training
}
