% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{kartsim_control}
\alias{kartsim_control}
\title{Controls Simulation}
\usage{
kartsim_control(direction, width = 32, height = 32, circuit = 1)
}
\arguments{
\item{direction}{A function \code{function(image, direction) {}} that returns
a direction as string with valid values: \code{"left"},  \code{"forward"} or 
\code{"right"} and which can make use of a raw png \code{image}.}

\item{width}{Width of captured image.}

\item{height}{Height of captured image.}

\item{circuit}{The circuit index, valid values: 1 or 2.

#' @examples 

library(kartsim)
if (interactive()) {
  kartsim_control(function(image) "right")
}}
}
\description{
Creates an ShinyGadget to control kart in simulation.
}
