% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{kartsim_play}
\alias{kartsim_play}
\title{Play Simulation}
\usage{
kartsim_play(width = NULL, height = NULL, circuit = 1)
}
\arguments{
\item{width}{While capturing, width of capture image.}

\item{height}{While capturing, height of capture image.}

\item{circuit}{The circuit index, valid values: 1 or 2.}
}
\description{
Creates an HTMLWidget to play the simulation.
}
\examples{

library(kartsim)
kartsim_play()

}
