% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{kartsim_shiny_output}
\alias{kartsim_shiny_output}
\title{Shiny Widget Output}
\usage{
kartsim_shiny_output(outputId, width = "100\%", height = "100\%")
}
\arguments{
\item{outputId}{The identifier for this widget.}

\item{width}{The width for this widget.}

\item{height}{The height for this widget.}
}
\description{
Provides a Shiny Widget for Output.
}
\examples{
library(shiny)

ui <- fluidPage(
  tags$head(
    tags$style(HTML("
      #kartsim {
        height: 270px !important;
        margin-top: 10px;
      }
    "))
  ),
  kartsim_shiny_output("kartsim"),
  textOutput("direction")
)

server <- function(input, output) {
kartsim_shiny_render(
  kartsim_play(size)
)

observeEvent(input$done, {
  input$direction <- input$direction
})
}

if (interactive()) {
  shinyApp(ui = ui, server = server)
}

}
