% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{kartsim_shiny_render}
\alias{kartsim_shiny_render}
\title{Shiny Widget Render}
\usage{
kartsim_shiny_render(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{The \code{expr} for \code{shinyRenderWidget}.}

\item{env}{The \code{env} for \code{shinyRenderWidget}.}

\item{quoted}{The \code{quoted} for \code{shinyRenderWidget}.}
}
\description{
Renders the Shiny Widget.
}
\seealso{
[kartsim_shiny_output()] for an example of using this function
within a 'Shiny' application.
}
